requirements.txt
pocketbase
PyJWT
cryptography
httpx[http2]
dotenv
requests